pyproject.toml
[project]
name = "voiceforge"
version = "0.1.0"
description = "VoiceForge - Qwen3 本地語音轉錄與智慧潤稿工具"
readme = "readme.md"
license = { text = "MIT" }
requires-python = ">=3.11,<3.13"
authors = [{ name = "VoiceForge Team" }]
keywords = ["asr", "qwen3", "voice", "transcription", "speech-to-text"]

dependencies = [
    "qwen-asr>=0.0.6",
    "huggingface-hub>=0.25.0",
    "customtkinter>=5.2.0",
    "sounddevice>=0.5.0",
    "numpy>=1.26.0",
    "pydub>=0.25.1",
    "pyperclip>=1.9.0",
    "keyboard>=0.13.5",
    "pyautogui>=0.9.54",
    "torch>=2.1.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "soundfile>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6.0",
]
vllm = [
    # vllm and the qwen-asr[vllm] extra require protobuf >=6 which is incompatible
    # with some versions of google-generativeai. Keep this optional and install
    # explicitly if you need local vLLM acceleration.
    "vllm==0.14.0",
]

# Google Generative AI client is optional because it pins protobuf ranges that
# conflict with vllm (installed via the vllm optional extras). Install only
# when you need Google GenAI features.
google = [
    "google-generativeai>=0.8.0",
]

[project.scripts]
voiceforge = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["core", "pages", "utils"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# google-generativeai requires protobuf <6; vllm requires protobuf >=6.30.
# Declare them as mutually exclusive so uv resolves each extra independently.
conflicts = [
    [
        { extra = "google" },
        { extra = "vllm" },
    ],
]

# Always fetch torch from the PyTorch CUDA 12.8 index so uv sync never
# downgrades to the CPU-only build from PyPI.
[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
